% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linereg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{dep}}{matrix.}

\item{\code{indep}}{matrix.}

\item{\code{betahat}}{matrix.}

\item{\code{predicted}}{matrix.}

\item{\code{residuals}}{matrix.}

\item{\code{inv_sqr_dep}}{matrix.}

\item{\code{deg_fre}}{numeric.}

\item{\code{sigmasq}}{numeric.}

\item{\code{var_betas}}{numeric.}

\item{\code{t_values}}{numeric.}

\item{\code{p_values}}{numeric.}

\item{\code{data}}{ANY.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Method coef}

\item{\code{model(y, x, ...)}}{Formula for reggression}

\item{\code{plot()}}{plots the residuals}

\item{\code{pred()}}{Method for prediction}

\item{\code{tvalue()}}{method t-value}
}}

